requirements.txt
allure-behave==2.13.2
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
behave==1.2.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
docopt==0.6.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
parse==1.19.1
parse-type==0.6.2
pipreqs==0.4.13
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.3
requests==2.31.0
selenium==4.14.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
yarg==0.1.9
